开始修改问题1.0

+++ app/src/main/java/com/ruanyun/australianews/model/VIPSouSuoInfo.java
@@ -110,7 +110,7 @@ public class VIPSouSuoInfo {
         private String outUrl;
         private String videoUrl;
         private String linkUrl;
-        private int normalPriceusd;
+        private double normalPriceusd;
         private int contentType;
         private double specialOffercny;
         private int priceUsd;
@@ -288,7 +288,7 @@ public class VIPSouSuoInfo {
             return linkUrl;
         }
 
-        public int getNormalPriceusd() {
+        public double getNormalPriceusd() {
             return normalPriceusd;
         }
